requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
argon2-cffi
python-dotenv
sqlalchemy